+++ src/components/pages/apps/matcalc/history/history-provider/plot.ts
@@ -1,3 +1,4 @@
+import { Axis } from '@/components/plot/axes/axis'
 import { IDBEntity } from '@/interfaces/db-entity'
 import { IClusterGroupRow } from '@/lib/cluster-group'
 
@@ -6,6 +7,11 @@ export interface IBasePlot extends IDBEntity {
   // groups to make plots so that they are independent
   // of history such that if user moves groups around
   // it won't affect any plots generated
+  axes?: {
+    x?: Axis | undefined
+    y?: Axis | undefined
+    colorbar?: Axis | undefined
+  }
   groupRows: IClusterGroupRow[]
   actions: string[]
   type: 'plot'